new toys

+++ bench.c
@@ -4,7 +4,8 @@
 static double now_ns(void) {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec*1e9 + ts.tv_nsec;
+    return (double)ts.tv_sec * 1e9
+         + (double)ts.tv_nsec;
 }
 
 double bench(double const goal_ns, double (*fn)(int loops)) {
